Add TaskHook.dll binary and source
The binary comes from the EventGhost 0.5.0-rc4 release.
The source directory comes from https://github.com/EventGhost/Extensions
and the most recent commit at the time I copied it was e031339.
As of this writing I don't know how to build the DLL from the source.
But I'll figure it out!