Multiple changes
Some code formatting
Adds import of win32con
Changed WindowInfo.title to a property that gets the window title each time it is called. due to the dynamic nature of window titles
Added window title change event. This is useful for windows that have tabs and when the tab changes it changes the title of the window. examples are web browsers
Changes Destroy to use postmessage due to permissions
Added Close method
Bugfix for all methods involving setting of the size or position
Changes repr to use process name
Moves the registration of the window notifications until after all self attributes are set. there is a slight change that a window message could come in and an attribute error would get raised.
Changes default parameters for Flash better speed and changed the number of times to 10 6 seemed to little.

Changed how SendKeystrokes functioned. it will get the top window for the process and send the keystrokes to that.

Changes setting of size and position not to change the z order of the windows